Move the async code back to overwrite the sync code
