+++ iotilegateway/iotilegateway/supervisor/__init__.py
@@ -1,5 +1,4 @@
 from .status_client import ServiceStatusClient
-from .async_status_client import AsyncServiceStatusClient
 from .supervisor import IOTileSupervisor
 
-__all__ = ['AsyncServiceStatusClient', 'ServiceStatusClient', 'IOTileSupervisor']
+__all__ = ['ServiceStatusClient', 'IOTileSupervisor']
